+++ sequence/topology/node.py
@@ -130,19 +130,21 @@ class Node(Entity):
 
         self.qchannels[another] = qchannel
 
-    def send_message(self, dst: str, msg: "Message", priority=inf) -> None:
+    def send_message(self, dst: str, msg: "Message", priority=inf, sender_delay: int = 0) -> None:
         """Method to send classical message.
 
         Args:
             dst (str): name of destination node for message.
             msg (Message): message to transmit.
             priority (int): priority for transmitted message (default inf).
+            sender_delay (int): sender-side delay (ps) before message is sent out on the channel (default 0).
+                                Include but not limited to processing delay, queueing delay, transmission delay, etc.
         """
         log.logger.debug(f"{self.name} send message {msg} to {dst}")
 
         if priority == inf:
             priority = self.timeline.schedule_counter
-        self.cchannels[dst].transmit(msg, self, priority)
+        self.cchannels[dst].transmit(msg, self, priority, sender_delay)
 
     def receive_message(self, src: str, msg: "Message") -> None:
         """Method to receive message from classical channel.
@@ -365,13 +367,15 @@ class QuantumRouter(Node):
                         protocol.received_message(src, msg)
                         break
 
-    def send_message(self, dst: str, msg: "Message", priority=inf) -> None:
+    def send_message(self, dst: str, msg: "Message", priority=inf, sender_delay: int = 0) -> None:
         """Method to send a classical message.
 
         Args:
             dst (str): name of the destination node to get the message.
             msg (Message): message to transmit.
             priority (int): priority for the transmitted message (default inf).
+            sender_delay (int): sender-side delay (ps) before message is sent out on the channel (default 0).
+                                Include but not limited to processing delay, queueing delay, transmission delay, etc.
         """
         if self.down:
             log.logger.debug(f"{self.name} is DOWN. Dropping message {msg} to {dst}")
@@ -380,7 +384,7 @@ class QuantumRouter(Node):
         log.logger.info(f"{self.name}: send message {msg} to {dst}")
         if priority == inf:
             priority = self.timeline.schedule_counter
-        self.cchannels[dst].transmit(msg, self, priority)
+        self.cchannels[dst].transmit(msg, self, priority, sender_delay)
 
     def set_down(self, down: bool):
         """Method to set the node status.
@@ -779,19 +783,21 @@ class ClassicalNode(ClassicalEntity):
 
         self.cchannels[another] = cchannel
 
-    def send_message(self, dst: str, msg: "Message", priority=inf) -> None:
+    def send_message(self, dst: str, msg: "Message", priority=inf, sender_delay: int = 0) -> None:
         """Method to send classical message.
 
         Args:
             dst (str): name of destination node for message.
             msg (Message): message to transmit.
             priority (int): priority for transmitted message (default inf).
+            sender_delay (int): sender-side delay (ps) before message is sent out on the channel (default 0).
+                                Include but not limited to processing delay, queueing delay, transmission delay, etc.
         """
         log.logger.info(f"{self.name} send message {msg} to {dst}")
 
         if priority == inf:
             priority = self.timeline.schedule_counter
-        self.cchannels[dst].transmit(msg, self, priority)
+        self.cchannels[dst].transmit(msg, self, priority, sender_delay)
 
     def receive_message(self, src: str, msg: "Message") -> None:
         """Method to receive message from classical channel.